refactor: 移动runner_logging
